SearchFacility diffusable et non diffusable
