+++ app/controllers/companies_controller.rb
@@ -8,7 +8,7 @@ class CompaniesController < ApplicationController
       redirect_to show_with_siret_companies_path(params[:siret], solicitation: @current_solicitation&.id)
     else
       # save_search(search_params[:query])
-      result = SearchFacility.new(search_params).from_full_text_or_siren if search_params.present?
+      result = SearchFacility::Diffusable.new(search_params).from_full_text_or_siren if search_params.present?
       respond_to do |format|
         format.html do
           if result.present?
